refactor: switch to binary encoding (uint16, little-endian)
Replace JSON serialization with a compact binary layout. All counters
use uint16 and dates are stored as uint16 days since 1980-01-01.

Recap:

The crawler used to keep a permanent bare clone of every repo
to compute the vitality index. With hundreds of repos that's
tens of GB of data.

Replace the mirrors with a shallow clone in a temp dir plus a small
versioned binary vitality.json per repo, holding only the
aggregated metadata the index needs. Subsequent runs reuse it via
--shallow-since=lastUpdated.

This is actually really cool and I'm very proud of myself, go me!

+++ git/repo_activity.go
@@ -61,7 +61,7 @@ func CalculateRepoActivity(repository common.Repository, days int, now time.Time
 	for i := range days {
 		cutoff := now.AddDate(0, 0, -i)
 
-		authorSet := map[uint8]struct{}{}
+		authorSet := map[uint16]struct{}{}
 		var commits, merges float64
 
 		cur := cache.FirstEntryDate
